added admin-side frontend, bug fix across controllers and services for the backend, fixed id parsing for individual posts and comments.

+++ backend/src/services/commentService.js
@@ -2,23 +2,29 @@ import prisma from "../config/prisma.js";
 
 const commentService = {
     getAllComments: async (id) => {
+        const postId = parseInt(id, 10);
+        if (isNaN(postId)) throw new Error("Invalid Post ID provided");
         return await prisma.comment.findMany({
             where: {
-                postId: id,
+                postId: postId,
             },
         });
     },
     postComment: async (id, body) => {
         const { author, message } = body;
-        await prisma.comment.create({
+        const postId = parseInt(id, 10);
+        if (isNaN(postId)) throw new Error("Invalid Post ID provided");
+        return await prisma.comment.create({
             data: {
                 author: author,
                 message: message,
-                postId: id,
+                postId: postId,
             },
         });
     },
-    deleteComment: async (id) => {
+    deleteComment: async (commentId) => {
+        const id = parseInt(commentId, 10);
+        if (isNaN(id)) throw new Error("Invalid Post ID provided");
         await prisma.comment.delete({
             where: {
                 id: id,
